Moved constants to API!
